requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
rapidfuzz>=3.0.0

# Logging
structlog>=23.2.0
rich>=13.7.0

# Testing
pytest>=7.4.3
pytest-mock>=3.12.0
responses>=0.24.1

# Type checking
mypy>=1.7.1
types-requests>=2.31.0

# Code quality
black>=23.12.0
ruff>=0.1.8
